output/dnssim: create internal-only header file
