+++ src/Makefile.am
@@ -42,7 +42,7 @@ dnsjit_LDADD = $(PTHREAD_LIBS) $(luajit_LIBS)
 
 # C source and headers
 dnsjit_SOURCES += core/thread.c core/compat.c core/channel.c core/object/null.c core/object/icmp.c core/object/ip.c core/object/udp.c core/object/ieee802.c core/object/gre.c core/object/pcap.c core/object/dns.c core/object/linuxsll.c core/object/ether.c core/object/payload.c core/object/loop.c core/object/icmp6.c core/object/tcp.c core/object/ip6.c core/receiver.c core/producer.c core/object.c core/log.c lib/clock.c input/mmpcap.c input/zero.c input/pcap.c input/fpcap.c filter/timing.c filter/split.c filter/ipsplit.c filter/copy.c filter/layer.c output/null.c output/tlscli.c output/respdiff.c output/pcap.c output/dnssim.c output/tcpcli.c output/dnscli.c output/udpcli.c
-dist_dnsjit_SOURCES += core/log.h core/producer.h core/assert.h core/compat.h core/object/udp.h core/object/payload.h core/object/gre.h core/object/icmp.h core/object/ip.h core/object/pcap.h core/object/dns.h core/object/loop.h core/object/ieee802.h core/object/ether.h core/object/linuxsll.h core/object/ip6.h core/object/icmp6.h core/object/tcp.h core/object/null.h core/object.h core/receiver.h core/channel.h core/timespec.h core/thread.h lib/clock.h input/zero.h input/fpcap.h input/pcap.h input/mmpcap.h filter/copy.h filter/layer.h filter/ipsplit.h filter/split.h filter/timing.h output/dnssim.h output/dnscli.h output/dnssim_ll.h output/pcap.h output/respdiff.h output/udpcli.h output/tlscli.h output/tcpcli.h output/null.h
+dist_dnsjit_SOURCES += core/log.h core/producer.h core/assert.h core/compat.h core/object/udp.h core/object/payload.h core/object/gre.h core/object/icmp.h core/object/ip.h core/object/pcap.h core/object/dns.h core/object/loop.h core/object/ieee802.h core/object/ether.h core/object/linuxsll.h core/object/ip6.h core/object/icmp6.h core/object/tcp.h core/object/null.h core/object.h core/receiver.h core/channel.h core/timespec.h core/thread.h lib/clock.h input/zero.h input/fpcap.h input/pcap.h input/mmpcap.h filter/copy.h filter/layer.h filter/ipsplit.h filter/split.h filter/timing.h output/dnssim.h output/dnscli.h output/dnssim_ll.h output/pcap.h output/respdiff.h output/udpcli.h output/tlscli.h output/tcpcli.h output/dnssim_internal.h output/null.h
 
 # Lua headers
 dist_dnsjit_SOURCES += core/timespec.hh core/object.hh core/channel.hh core/receiver.hh core/producer.hh core/object/icmp.hh core/object/ether.hh core/object/pcap.hh core/object/loop.hh core/object/dns.hh core/object/ip.hh core/object/null.hh core/object/icmp6.hh core/object/udp.hh core/object/ieee802.hh core/object/ip6.hh core/object/gre.hh core/object/linuxsll.hh core/object/tcp.hh core/object/payload.hh core/log.hh core/thread.hh lib/clock.hh input/mmpcap.hh input/zero.hh input/pcap.hh input/fpcap.hh filter/split.hh filter/copy.hh filter/ipsplit.hh filter/timing.hh filter/layer.hh output/udpcli.hh output/dnscli.hh output/pcap.hh output/null.hh output/respdiff.hh output/tlscli.hh output/dnssim.hh output/tcpcli.hh
