+++ src/output/dnssim.c
@@ -22,163 +22,7 @@
 
 #include "output/dnssim.h"
 #include "output/dnssim_ll.h"
-#include "core/assert.h"
-#include "core/object/ip.h"
-#include "core/object/ip6.h"
-#include "core/object/payload.h"
-#include "core/object/dns.h"
 
-typedef struct _output_dnssim_s _output_dnssim_t;
-typedef struct _output_dnssim_client_s _output_dnssim_client_t;
-typedef struct _output_dnssim_request_s _output_dnssim_request_t;
-typedef struct _output_dnssim_connection_s _output_dnssim_connection_t;
-typedef struct _output_dnssim_query_s _output_dnssim_query_t;
-typedef struct _output_dnssim_query_udp_s _output_dnssim_query_udp_t;
-typedef struct _output_dnssim_query_tcp_s _output_dnssim_query_tcp_t;
-typedef struct _output_dnssim_source_s _output_dnssim_source_t;
-
-struct _output_dnssim_source_s {
-    _output_dnssim_source_t* next;
-    struct sockaddr_storage addr;
-};
-
-typedef enum _output_dnssim_read_state {
-    _OUTPUT_DNSSIM_READ_STATE_CLEAN,
-    _OUTPUT_DNSSIM_READ_STATE_DNSLEN,
-    _OUTPUT_DNSSIM_READ_STATE_DNSMSG,
-    _OUTPUT_DNSSIM_READ_STATE_INVALID
-} _output_dnssim_read_state_t;
-
-struct _output_dnssim_connection_s {
-    _output_dnssim_connection_t* next;
-
-    uv_tcp_t handle;
-    uv_connect_t conn_req;
-    uv_timer_t timeout;
-
-    /* List of queries that have been queued (pending write callback). */
-    _output_dnssim_query_t* queued;
-
-    /* List of queries that have been sent over this connection. */
-    _output_dnssim_query_t* sent;
-
-    _output_dnssim_client_t* client;
-
-    enum {
-        _OUTPUT_DNSSIM_CONN_INITIALIZED,
-        _OUTPUT_DNSSIM_CONN_CONNECTING,
-        // _OUTPUT_DNSSIM_CONN_CONNECTED,
-        _OUTPUT_DNSSIM_CONN_ACTIVE,
-        _OUTPUT_DNSSIM_CONN_CLOSING,
-        _OUTPUT_DNSSIM_CONN_CLOSED
-    } state;
-
-    _output_dnssim_read_state_t read_state;
-
-    /* Total length of the expected stream data (either 2 for dnslen, or dnslen itself). */
-    size_t recv_len;
-
-    /* Current position in the receive buffer. */
-    size_t recv_pos;
-
-    /* Receive buffer used for incomplete messages or dnslen. */
-    char* recv_data;
-    bool recv_free_after_use;
-};
-
-struct _output_dnssim_client_s {
-    /* List of connections.
-     * Multiple connections may be used (e.g. some are already closed for writing).
-     */
-    _output_dnssim_connection_t* conn;
-
-    /* List of queries that are pending to be sent over any available connection. */
-    _output_dnssim_query_t* pending;
-};
-
-struct _output_dnssim_s {
-    output_dnssim_t pub;
-
-    output_dnssim_transport_t transport;
-    uv_loop_t loop;
-    struct sockaddr_storage target;
-    _output_dnssim_source_t* source;
-
-    _output_dnssim_client_t* client_arr;
-
-    uv_timer_t stats_timer;
-
-    void (*create_request)(output_dnssim_t*, _output_dnssim_client_t*,
-        core_object_payload_t*);
-};
-
-struct _output_dnssim_query_s {
-    _output_dnssim_query_t* next;
-    output_dnssim_transport_t transport;
-    _output_dnssim_request_t* req;
-
-    /* Query state, currently used only for TCP. */
-    enum {
-        _OUTPUT_DNSSIM_QUERY_PENDING_WRITE,
-        _OUTPUT_DNSSIM_QUERY_PENDING_WRITE_CB,
-        _OUTPUT_DNSSIM_QUERY_SENT,
-        _OUTPUT_DNSSIM_QUERY_CLOSED
-    } state;
-};
-
-struct _output_dnssim_query_udp_s {
-    _output_dnssim_query_t qry;
-    uv_udp_t* handle;
-    uv_buf_t buf;
-    //uv_timer_t* qry_retransmit;
-};
-
-struct _output_dnssim_query_tcp_s {
-    _output_dnssim_query_t qry;
-    _output_dnssim_connection_t* conn;
-
-    uv_write_t write_req;
-    uv_buf_t bufs[2];
-};
-
-struct _output_dnssim_request_s {
-    _output_dnssim_query_t* qry;
-    _output_dnssim_client_t* client;
-    core_object_payload_t* payload;
-    core_object_dns_t* dns_q;
-    uint64_t created_at;
-    uint64_t ended_at;
-    uv_timer_t* timeout;
-    uint8_t timeout_closing;
-    uint8_t ongoing;
-    // enum {
-    //     _OUTPUT_DNSSIM_
-    output_dnssim_t* dnssim;
-};
-
-static core_log_t _log = LOG_T_INIT("output.dnssim");
-static output_dnssim_t _defaults = {
-    LOG_T_INIT_OBJ("output.dnssim"),
-    0, 0, 0,
-    NULL, NULL, NULL,
-    0, 0,
-    2000, 0
-};
-
-static output_dnssim_stats_t _stats_defaults = {
-    NULL, NULL,
-    NULL,
-    0, 0,
-    0, 0, 0,
-    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
-};
-
-// forward declarations
-static void _close_query_udp(_output_dnssim_query_udp_t* qry);
-static void _close_query_tcp(_output_dnssim_query_tcp_t* qry);
-static void _close_request_timeout_cb(uv_handle_t* handle);
-static void _close_request_timeout(uv_timer_t* handle);
-static void _close_tcp_connection(_output_dnssim_connection_t* conn);
 
 uint64_t _now_ms()
 {
